Network Policy Advisor: move advisor in separate package
This is to avoid linking too many packages in the kubectl-gadget CLI
tool.

Internal imports before this patch:
```
cmd/kubectl-gadget
    cmd/kubectl-gadget/utils
        pkg/api/v1alpha1
        pkg/factory
        pkg/k8sutil
        pkg/types
    pkg/gadgets/dns/types
    pkg/gadgets/networkpolicy
        pkg/gadgets
            pkg/container-collection
                pkg/gadgettracermanager/api
                pkg/gadgettracermanager/containerutils
                pkg/gadgettracermanager/pubsub
                pkg/runcfanotify
        pkg/gadgets/networkpolicy/types
    pkg/gadgets/socket-collector/types
    pkg/resources
```
Internal imports after this patch:
```
cmd/kubectl-gadget
    cmd/kubectl-gadget/utils
        pkg/api/v1alpha1
        pkg/factory
        pkg/k8sutil
        pkg/types
    pkg/gadgets/dns/types
    pkg/gadgets/networkpolicy/advisor
        pkg/gadgets/networkpolicy/types
    pkg/gadgets/socket-collector/types
    pkg/resources
```
