+++ pkg/gadgets/networkpolicy/advisor/advisor_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package networkpolicy
+package advisor
 
 import (
 	"io/ioutil"
